raspberry/raspberry_requirements.txt
# 라즈베리파이 필수 패키지 리스트
# Python 3.8+ 필요

# 기본 이미지 처리
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# 딥러닝 (라즈베리파이용 경량 버전)
torch==2.0.0  # CPU 버전
torchvision==0.15.0
# ARM 아키텍처용 설치:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu

# ONNX Runtime (선택사항 - 더 빠른 추론)
onnxruntime==1.16.0
# ARM용: pip install onnxruntime-arm64

# 네트워크 통신
aiohttp==3.8.5
requests==2.31.0
asyncio  # Python 내장

# 데이터 처리
scikit-learn==1.3.0  # cosine_similarity용
scipy==1.11.2

# 로깅 및 유틸리티
python-dateutil==2.8.2

# =====================================
# 선택사항 (성능 최적화)
# =====================================
# TensorFlow Lite (더 가벼운 대안)
# tflite-runtime==2.13.0

# Edge TPU 사용 시
# pycoral~=2.0.0
# tflite-runtime==2.5.0.post1